requirements.txt
# Webtop Homework Skill Requirements

# Core dependencies
requests>=2.25.1
google-auth>=1.6.3
google-auth-oauthlib>=0.4.1
google-api-python-client>=2.0.0
pickle5>=0.0.11

# Web automation
playwright>=1.0.0
beautifulsoup4>=4.9.3
selenium>=3.141.0

# Data handling
jsonschema>=3.2.0
python-dateutil>=2.8.1

# Messaging
whatsapp-bot>=0.1.0  # For clawdbot integration

# Utility
python-dotenv>=0.15.0
tqdm>=4.50.0
click>=7.1.2

# Development (optional)
pytest>=6.2.0
black>=21.0.0
flake8>=3.8.0